# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "En esta documentación se encuentran los detalles de los endpoints presentes en el artefacto BFF del proyecto Kümelen",
        "title": "Documentación Artefacto BFF",
        "contact": {
            "name": "Diego Sepúlveda",
            "email": "diego.sepulvedas@utem.cl"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/auth/login": {
            "post": {
                "description": "Permite a los estudiantes iniciar sesión con sus credenciales de Pasaporte.UTEM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Autenticación"
                ],
                "summary": "Inicio de Sesión",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Student"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "BadGateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/auth/refresh-token": {
            "post": {
                "description": "Para reducir la cantidad de veces que un usuario debe iniciar sesión, se diponibiliza un endpoint para actualizar el token de consultas a los diferentes servicios",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Autenticación"
                ],
                "summary": "Actualización de Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "BadGateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/auth/validate-token": {
            "post": {
                "description": "Con el fin de detectar la validez de un token se disponibiliza este endpoint para verificar su integridad",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Autenticación"
                ],
                "summary": "Validación de Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/catalog": {
            "get": {
                "description": "Devuelve el contenido a mostrar en la pantalla de catélogo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Catálogo"
                ],
                "summary": "Pantalla de Catálogo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Token-Autorización",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CatalogPage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "BadGateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/category/{id}/posts": {
            "get": {
                "description": "Devuelve todas las publicaciones que presentan la categoría enviada por parámetro",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Categorías"
                ],
                "summary": "Publicaciones de una categoría",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Token-Autorización",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "123456",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CategoryPostsPage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "BadGateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/home": {
            "get": {
                "description": "Devuelve el contenido a mostrar en la pantalla de inicio de la aplicación",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Pantalla Principal"
                ],
                "summary": "Pantalla de Inicio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Token-Autorización",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HomePage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "BadGateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/post/search/{searchTerm}": {
            "get": {
                "description": "Devuelve todas las publicaciones que cumplen con el término de búsqueda",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Publicaciones"
                ],
                "summary": "Resultados de Búsqueda",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Token-Autorización",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "bienestar",
                        "name": "searchTerm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PostSearchResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "BadGateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/post/{id}": {
            "get": {
                "description": "Devuelve el detalle de una publicación pora ser mostrado en la aplicación",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Publicaciones"
                ],
                "summary": "Pantalla de Publicación",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Token-Autorización",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "123456",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PostPage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "BadGateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/service/{id}": {
            "get": {
                "description": "Devuelve el detalle de un servicio estudiantil para visualizarlo en la aplicación",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Servicios Estudiantiles"
                ],
                "summary": "Pantalla de Servicio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Token-Autorización",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "123456",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ServicePage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "BadGateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/service/{id}/posts": {
            "get": {
                "description": "Devuelve todas las publicaciones creadas por un servicio estudiantil especificado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BFF V1 - Servicios Estudiantiles"
                ],
                "summary": "Publicaciones de un Servicio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Token-Autorización",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "123456",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ServicePostsPage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "BadGateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CatalogPage": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Category"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                }
            }
        },
        "models.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CategoryPostsPage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "posts_for_preview": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostForPreview"
                    }
                }
            }
        },
        "models.Contact": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.HomePage": {
            "type": "object",
            "properties": {
                "services_with_posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ServiceWithPosts"
                    }
                }
            }
        },
        "models.Person": {
            "type": "object",
            "properties": {
                "charge": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profile_picture_url": {
                    "type": "string"
                }
            }
        },
        "models.PostCategory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.PostForPreview": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.PostPage": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostCategory"
                    }
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.PostSearchResult": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostForPreview"
                    }
                }
            }
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.ServicePage": {
            "type": "object",
            "properties": {
                "contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Contact"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "persons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Person"
                    }
                }
            }
        },
        "models.ServicePostsPage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "posts_for_preview": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostForPreview"
                    }
                }
            }
        },
        "models.ServiceWithPosts": {
            "type": "object",
            "properties": {
                "abbreviation": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "posts_for_preview": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostForPreview"
                    }
                }
            }
        },
        "models.Student": {
            "type": "object",
            "properties": {
                "career": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    }
}